# app/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.4.0"
  title: URJC gestiongrupos Backend PWA
# during dev, should point to your local machine
host: localhost:10081
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/octet-stream
# format of the responses to the client (Accepts)
produces:
  - application/json
  - application/xml
paths:
...
  /private/calls: 
    x-swagger-router-controller: call.controller
    get: 
      tags: 
        - "Calls" 
      description: Get calls
      operationId: getCalls
      parameters: 
        - in: header 
          name: authorization 
          description: Security Token 
          required: true 
          type: string 
      responses: 
        "200": 
          description: Success 
          schema: 
            $ref: "#/definitions/CallsDataListResponse"
          headers:
            authorization:
              description: Security Token
              type: string
          examples:
            application/json: {
                "id": "call-ucvb83kzsjw0pfsep",
                "name": "test call 1",
                "company": "URJC-Gestiongrupos",
                "initDate": "2019-06-03T00:00:00.000Z",
                "endDate": "2019-09-15T00:00:00.000Z",
                "currentDate": "2019-09-16T00:00:00.000Z",
                "agenda": [{"title": "archive_ucvb83kzsjw0pfsep.pdf", "data": "<<BASE64>>"}],
                "organizers": [
                  {"name": "Ana", "position": "Presidenta"}, 
                  {"name": "Juan Carlos", "position": "Secretario"},
                ],
                "totalShares": 200000,
	              "minShares": 120000,
                "options": ["A favor", "En contra", "Abstención", "En blanco"],
	   	          "participants": [
	   		          {"id": "11a", "name": "Cristina Castaño", "shares": 200},
                  {"id": "22b", "name": "Ramón González", "shares": 100}, 
                 ],
                "counting": [],
                "currentResult": [],
          }
        "401":
          description: Not authenticated in the system
          schema:
            $ref: "#/definitions/UnauthenticatedResponse"
          examples:
            error:
              code: 401
              message: Security header not received
              description: Security header not received in [Auth Middleware]:authenticate
        "403":
          description: Not authorized to execute this method
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
          examples:
            error:
              code: 403
              message: User not authorized
              description: User not authorized in [Call Controller]:getCalls
        "404":
          description: Calls not found
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 404
              message: No calls found
              description: No calls found in [Call Controller]:getCalls
        default: 
          description: Error 
          schema: 
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 500
              message: Internal Server Error
              description: Internal Server Error in [Call Controller]:getCalls
  /private/calls/{idCall}: 
    x-swagger-router-controller: call.controller
    get: 
      tags: 
        - "Calls" 
      description: Get call by Id
      operationId: getCallById
      parameters: 
        - in: header 
          name: authorization 
          description: Security Token 
          required: true 
          type: string
        - in: path
          name: idCall
          description: idCall
          required: true 
          type: string 
      responses: 
        "200": 
          description: Success 
          schema: 
            $ref: "#/definitions/CallDataResponse"
          headers:
            authorization:
              description: Security Token
              type: string
          examples:
            application/json: {
                "id": "call-ucvb83kzsjw0pfsep",
                "name": "test call 1",
                "company": "URJC-Gestiongrupos",
                "initDate": "2019-06-03T00:00:00.000Z",
                "endDate": "2019-09-15T00:00:00.000Z",
                "currentDate": "2019-09-16T00:00:00.000Z",
                "agenda": [{"title": "archive_ucvb83kzsjw0pfsep.pdf", "data": "<<BASE64>>"}],
                "organizers": [
                  {"name": "Ana", "position": "Presidenta"}, 
                  {"name": "Juan Carlos", "position": "Secretario"},
                ],
                "totalShares": 200000,
	              "minShares": 120000,
                "options": ["A favor", "En contra", "Abstención", "En blanco"],
	   	          "participants": [
	   		          {"id": "11a", "name": "Cristina Castaño", "shares": 200},
                  {"id": "22b", "name": "Ramón González", "shares": 100}, 
                 ],
                "counting": [],
                "currentResult": [],
          }
        "401":
          description: Not authenticated in the system
          schema:
            $ref: "#/definitions/UnauthenticatedResponse"
          examples:
            error:
              code: 401
              message: Security header not received
              description: Security header not received in [Auth Middleware]:authenticate
        "403":
          description: Not authorized to execute this method
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
          examples:
            error:
              code: 403
              message: User not authorized
              description: User not authorized in [Call Controller]:getCallById
        "404":
          description: Call not found
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 404
              message: Call not found
              description: Call not found in [Call Controller]:getCallById
        default: 
          description: Error 
          schema: 
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 500
              message: Internal Server Error
              description: Internal Server Error in [Call Controller]:getCallById
  /private/calls/{idCall}/summary: 
    x-swagger-router-controller: call.controller
    get: 
      tags: 
        - "Calls" 
      description: get votes summary from a call
      operationId: getBallotResult
      parameters: 
        - in: header 
          name: authorization 
          description: Security Token 
          required: true 
          type: string 
        - in: path
          name: idCall
          description: idCall
          required: true 
          type: string 
      responses:        
        default: 
          description: Error 
          schema: 
            $ref: "#/definitions/ErrorResponse"
          examples:
            error:
              code: 500
              message: Internal Server Error
              description: Internal Server Error in [Call Controller]:getBallotResult

  # ##################################################################
  # Blocks
  # ##################################################################
...
definitions:
...
  InnerErrorResponse:
    type: object
    properties:
      code:
        type: number
      message:
        type: string
      description:
        type: string
  ErrorResponse:
    type: object
    properties:
      error:
        $ref: "#/definitions/InnerErrorResponse"
        
  # Authentication Responses
...
  UnauthenticatedResponse:
    type: object
    properties:
      error:
        $ref: "#/definitions/InnerErrorResponse"

  # Authorization Responses
  UnauthorizedResponse:
    type: object
    properties:
      error:
        $ref: "#/definitions/InnerErrorResponse"

  # Vote Responses
...
  ImageObject:
    type: object
    properties:
      title:
        type: string
      data:
        type: string
        format: byte
  NamePositionObject:
    type: object
    properties:
      name:
        type: string
      position:
        type: string
  IdDescriptionObject:
    type: object
    properties:
      id:
        type: string
      description:
        type: string
  IdNameSharesObject:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      shares:
        type: integer
  IdvotedIdparticipantOptionObject:
    type: object
    properties:
      idVoted:
        type: string
      idParticipant:
        type: string
      option:
        type: string
  IdvotedOptionSharesObject:
    type: object
    properties:
      idVoted:
        type: string
      option:
        type: string
      shares:
        type: integer
  CallDataResponse: 
    type: object      
    properties:
      id:
        type: string
      name: 
        type: string 
      company: 
        type: string
      initDate:
        type: string
        format: date-time
      endDate:
        type: string
        format: date-time
      currentDate:
        type: string
        format: date-time
      agenda: 
        type: array
        items:
          $ref: "#/definitions/ImageObject"
      organizers:
        type: array
        items:
          $ref: "#/definitions/NamePositionObject"
      totalShares:
        type: integer
      minShares:
        type: integer
      proposals:
        type: array
        items:
          $ref: "#/definitions/IdDescriptionObject"
      options:
        type: array
        items:
          type: string
      participants:
        type: array
        items: 
          $ref: "#/definitions/IdNameSharesObject"
      currentResult:
        type: array
        items:
          $ref: "#/definitions/IdvotedIdparticipantOptionObject"
      counting:
        type: array
        items:
          $ref: "#/definitions/IdvotedOptionSharesObject"
  CallsDataListResponse: 
    type: object
    properties:
      totalElements:
        type: number
      elements:
        type: array
        items: 
          $ref: "#/definitions/CallDataResponse"

  # Block Responses
...